colon_global/requirements.txt
torch~=1.6.0
torchvision==0.6.0
scikit-learn~=0.23.1
scikit-image~=0.17.2
pandas~=1.0.5
tqdm~=4.48.0
pillow~=7.1.2
numpy~=1.19.0
ipdb~=0.13.3
google-cloud-storage==1.28.1
Keras-Preprocessing~=1.1.2
matplotlib
scipy
pyyaml
easydict
tensorflow
termcolor
faiss-gpu
keras